graph visualization overhaul, cylinder graph
graphviz layouts can be specified, and there is an API for setting edge
and vertex attributes. This is intended to make it somewhat painless to
visualize cold-peer selection inside a jupyter notebook. The cylinder
graph is also relevant for that: our idea for an initial experiment is
to try a selection algorithm on a cylindrical topology, i.e. a crude
approximation of the Earth with a bunch of peers evenly spaced out
around it.