Generalise Driver interface over Maybe bytes to driver state
The runPeer no longer even knows about codecs, channels or even bytes,
the trailing data it passes around could be just a driver state.

So swap from Maybe bytes everywhere to a general driver state, where the
driverSimple of course uses Maybe bytes. To properly abstract this we
have to add a startDState to the driver.